removing/disabling some modules, added new functionality and package update
Added module(translate) and functionality to interact with YandexGPT.
Implemented the functionality of dialogs and their storage.
Terms module has been temporarily deactivated, as this functionality is not required for mvp.
Removed forward-text module due to different way of interacting with YandexGPT was implemented.
Updated npm packages.
Some minor fixes.